No error when re-cultivating
This is useful when cultivating an entire directory, which might contain
some cultivated and some non-cultivated files.

An error still occurs when the file is owned by a different package than
the one you are trying to cultivate to, since there is no good way to
resolve such a situation.

Also got rid of a superfluous lexists check.